readme: fix commit message guideline
For some time the `commit message` section stablished the commit titles' subject should begin with a capital letter. The example, though, had a minuscule letter at the beginning of the subject and, probably because of that, we started following that guideline instead.

This PR aims at standardizing that behaviour in order to have our commit style guide in accord to how we develop in house.

This is inteded to be looked at as a natural evolution of the guideline, not as a simple change to comply with a rule that wasn't being followed.